Minify wgsl
Target esnext
